use int64 for id instead int

+++ database/schema.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS users (
-    id int unsigned NOT NULL AUTO_INCREMENT,
+    id bigint unsigned NOT NULL AUTO_INCREMENT,
     name varchar(100) NOT NULL,
     email varchar(100) NOT NULL,
     password varchar(100) NOT NULL,
